+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+#### 0.1.1
+- Added Shadow DOM support (`event.target` returns the custom element, not the
+  actual focused element that is inside the shadow tree of that custom element)
+- Properly handled disabled targets (input and button)
+
 #### 0.1.0
 -	Initial version